pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "json2ubl"
version = "1.0.0"
description = "Production-grade JSON to UBL 2.1 XML converter with schema-driven mapping"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "SherozShaikh", email = "shaikh.sheroz07@gmail.com"}
]
keywords = ["ubl", "xml", "json", "invoice", "converter", "edi"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
    "Topic :: Office/Business :: Financial",
]

dependencies = [
    "lxml>=4.9.4,<5.0",
    "pydantic>=2.7.0,<3.0",
    "pyyaml>=6.0.1,<7.0",
    "loguru>=0.7.2,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9.0",
    "pytest-cov>=5.0.0,<6.0",
    "black>=24.10.0,<25.0",
    "ruff>=0.8.0,<1.0",
    "mypy>=1.13.0,<2.0",
]

[project.urls]
Homepage = "https://github.com/sherozshaikh/json2ubl"
Documentation = "https://github.com/sherozshaikh/json2ubl#readme"
Repository = "https://github.com/sherozshaikh/json2ubl.git"
Issues = "https://github.com/sherozshaikh/json2ubl/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
json2ubl = [
    "schemas/ubl-2.1/**/*.xsd",
    "config/*.yaml",
    "schemas/cache/*.json",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=json2ubl --cov-report=term --cov-report=html"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

include = '\.pyi?$'
extend-exclude = '''
/(
  \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "C",
    "B",
]

ignore = [
    "E501",
    "C901",
]
